Renames file to match class definition
