+++ ui/analyze_dialog.h
